Task two/requirements.txt
Django>=4.0
djangorestframework>=3.14
transformers>=4.35
torch>=2.0
requests>=2.31
